requirements.txt
asgiref==3.8.1
Django==5.2.3
django-widget-tweaks==1.5.0
python-dotenv==1.1.0
sqlparse==0.5.3
typing_extensions==4.14.0
